--- src/playwright/axe.ts
@@ -1,9 +1,0 @@
-export {
-  injectAxe,
-  configureAxe,
-  getAxeResults,
-  getViolations,
-  reportViolations,
-  checkA11y,
-  DefaultTerminalReporter,
-} from 'axe-playwright';
