+++ src/util/index.ts
@@ -1,17 +1,5 @@
-import type { Config } from '@jest/types';
-
 export * from './getCliOptions';
 export * from './getTestRunnerConfig';
 export * from './getStorybookMain';
 export * from './getStorybookMetadata';
 export * from './getParsedCliOptions';
-
-export {
-  toMatchImageSnapshot,
-  configureToMatchImageSnapshot,
-  updateSnapshotState,
-} from 'jest-image-snapshot';
-
-type JestConfig = Config.InitialOptions;
-
-export { type JestConfig };
